+++ FRONT/match-app-front/src/index.css
@@ -15,7 +15,7 @@ body{
   overflow-x: hidden;
   height: fit-content;
   max-width: 100vw;
-  background-color: black;
+  background-color: rgb(255, 243, 243);
   font-family: 'Exo 2', sans-serif;
 
 }
