+++ FRONT/match-app-front/src/index.js
@@ -1,8 +1,8 @@
 // src/index.js
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App'; // Certifique-se de que o caminho está correto
-import './index.css'; // Se você tiver um arquivo CSS
+import App from './App';
+import './index.css'; 
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
